+++ src/main/java/org/peter_lukas/shirtso/commercial/product/dto/ProductDto.java
@@ -13,7 +13,7 @@ public record ProductDto (
         String description,
         BigDecimal price,
         Currencies currency,
-        int imageId,
+        Long imageId,
         int subcategoryId,
         String supplier,
         long stock,
