+++ src/main/java/org/peter_lukas/shirtso/commercial/product/image/ProductImageController.java
@@ -39,7 +39,7 @@ public class ProductImageController {
 
     @GetMapping("/{imageId}")
     @LogExecutionTime
-    public ResponseEntity<ProductImage> getImageById(@PathVariable int imageId) {
+    public ResponseEntity<ProductImage> getImageById(@PathVariable Long imageId) {
         return productImageService.getImageById(imageId)
                 .map(ResponseEntity::ok)
                 .orElse(ResponseEntity.notFound().build());
@@ -49,7 +49,7 @@ public class ProductImageController {
     @LogExecutionTime
     @RolesAllowed(USER_WRITE)
     public ResponseEntity<ProductImage> updateImage(
-            @PathVariable int imageId,
+            @PathVariable Long imageId,
             @Valid @RequestBody CreateImageRequestDto request) {
         return ResponseEntity.ok(productImageService.updateImage(imageId, request));
     }
@@ -58,13 +58,13 @@ public class ProductImageController {
     @ResponseStatus(HttpStatus.NO_CONTENT)
     @LogExecutionTime
     @RolesAllowed(USER_WRITE)
-    public void deleteImage(@PathVariable int imageId) {
+    public void deleteImage(@PathVariable Long imageId) {
         productImageService.deleteImage(imageId);
     }
 
     @GetMapping("/{imageId}/products")
     @LogExecutionTime
-    public List<UUID> getProductsUsingImage(@PathVariable int imageId) {
+    public List<UUID> getProductsUsingImage(@PathVariable Long imageId) {
         return productImageService.getProductsUsingImage(imageId);
     }
 }
