+++ src/main/java/org/peter_lukas/shirtso/commercial/product/image/ProductImageRepository.java
@@ -2,5 +2,5 @@ package org.peter_lukas.shirtso.commercial.product.image;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductImageRepository extends JpaRepository<ProductImage, Integer> {
+public interface ProductImageRepository extends JpaRepository<ProductImage, Long> {
 }
