+++ src/main/java/org/peter_lukas/shirtso/commercial/product/image/dto/AssociateImageRequestDto.java
@@ -5,7 +5,7 @@ import jakarta.validation.constraints.NotNull;
 public record AssociateImageRequestDto(
 
         @NotNull(message = "Image ID can not be null")
-        Integer imageId,
+        Long imageId,
 
         boolean isPrimary,
