+++ src/main/java/org/peter_lukas/shirtso/commercial/product/image/dto/ProductImageDto.java
@@ -1,7 +1,7 @@
 package org.peter_lukas.shirtso.commercial.product.image.dto;
 
 public record ProductImageDto(
-        int imageId,
+        Long imageId,
         String imageUrl,
         String altText,
         boolean isPrimary,
